backend/requirements.txt
django
pymysql
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
drf-nested-routers
requests
openai
cryptography
